Separate the Test and Development Databases
Every time the tests run, the database would be reset, forcing me to re-create
a user. This is annoying.

I've added a .env.test for holding the .test environments settings; as well as
enhanced the database tasks.

Now to set up the test database you need to:

RACK_ENV=test rake db:migrate