Add Jasmine
Jasmine is kind of a pig to set up, but once you've got it rolling everything is amazing.

I wanted to ensure that:

1. I could load my projects javascript files
2. I could run the `rake jasmine` command and see a test pass.
3. I didn't include any of that lovely generated code in my git repo

The test here is not a very valuable test, but it allowed me to ensure Jasmine
was set up properly. I expect to delete it once I have some actual
behavior-focused tests.