Add jasmine-jquery and jasmine-fixture
These are two of the most useful libraries you can have in your jasmine
arsenal.

Jasmine-Jquery makes it easy to test against the DOM using jQuery. It comes
with dozens of matchers like 'toHaveClass' or 'toHandleWith' which allow you to
express your desired behavior with very little work

Jasmine-fixture makes it a snap to insert stuff into the DOM without doing a
ton of work. Your JavaScript is often going to be coupled to the DOM; so you'll
be doing a lot of tests which have to deal with your HTML structure.