Talk.save sends the talks attributes to the server
At this point, we're sending data to the server, not just making an
empty HTTP request.

We're choosing to send JSON(Javascript Object Notation). JSON is text which
represents the talks attributes object.

That's all JSON is; another way to store data in a way that both computers and
people can read.

Both Javascript and Ruby make it easy to convert JSON strings into hashes (in
rubies case) or objects (in javascripts case).

The notion of storing the same data in different forms is known as `encoding`.

Data a browser can display as a web page is encoded as HTML.
Data our music player can play as music is `encoded` as an MP3 (or any number
of competing encodings).

We could, if we so desired, even develop our own form of encoding data and
sending it forth to the server. This would be fun, but not terribly
practical.