Talks with ids are saved via a PUT to /talks/:id
Here I'm agressively extracting private methods, signified by the `__` or
'dunderscore'. I'm choosing not to test these private methods because they're
incidental to the behavior of the Talk object.

I'm choosing to expose `isPersisted`, because I can imagine it being useful to
outside collaborators. Some people will wait until another object actually
wants to call the isPersistd method befor publicly exposing it. This is also
valid.

I also refactored my test organization to focus more on the `inputs` (i.e. what
attributes the talk is instantiated with) than the individual methods. This is
somewhat controversial; as standard practice is to describe each method with
the permutations next to it.

Some of you may be thinking "man, it looks like the Talk object does too many
things. It's sending stuff to the server as well as holding the talk
attributes!

To you I say, well done. I concur. However, because there isn't a need to share
this behavior quite yet I will wait to extract an object which is responsible
for sending data to the server.