Convert createTalkView to TalksView
Technically the talkView contains more than one talk. I'm renaming it to
TalksView to be more accurate.

Also, I'm converting it to use a constructor function so it's consistent with
the Talk model.

You'll notice I'm using the instncaeof keyword to test that Thunder's talksView
object is actually a TalksView. This is reasonable only because I've thoroughly
tested TalksView independently.

The Thunder object's one job is to collect the different javascript components.