save is called on the new talk
Our tests are getting pretty crazy-complex. This is because the behavior we're
testing is actually quite complex as well.

We're decomposing the interactions between a computer that could be anywhere on
the Internet and our web application server; and we're accounting for that
being completely asynchronous.

This is also why most people don't test their view logic. It's. Really. Hard.