Append the saved talk to the DOM
Now we're beginning to get the meat. Most of the tests we've been writing so
far are tightly coupled to a specific implementation.

We're leveraging the fact that the Talk.save method returns a deferred object
so we can choose whether the save succeeds or fails in the test itself; and
then assert we're handling that properly.