Append errors when save talk fails
Here we begin to see that `this` sometimes gets changed.

`this` is a keyword which means `the object that the function
was called upon`

The object the function is called on is known as the object's `context`

```
function exploreThis() {
  console.log(this);
}

exploreThis();
// Prints the window, since we're not explicitely calling exploreThis on any
// object
```

We can also put the exploreThis function on an object; which will change what
`this` is when exploreThis is called.

```

coolObject = {
  exploreThisOnAnObject: exploreThis
}

coolObject.exploreThisOnAnObject();
// Prints the `coolObject` object.

```

But wait, there's more!

We can actually call `exploreThis` and tell it what `this` should be.

```
notCoolObject = {
  attributes: "I'm not as cool as the coolObject."
}

exploreThis.call(notCoolObject)
// Prints the notCoolObject
```

`context` and `this` are a huge part of javascript's strength, but also one of
the hardest things to grasp coming from ruby.